Mod: modularizes task input form.
