+++ src/App.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { useState } from 'react';
+import TaskInputForm from './components/TaskInputForm'
+import TaskList from './components/TaskList'
 
 type Task = {
   id: number,
@@ -17,37 +19,11 @@ const initialState: Task[] = [
 
 const App: React.FC = () => {
   const [ tasks, setTasks ] = useState(initialState)
-  const [ inputTitle, setInputTitle ] = useState('')
-  const [ count, setCount ] = useState(tasks.length + 1)
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setInputTitle(e.target.value)
-  }
-
-  const handleSubmit = () => {
-    setCount(count + 1)
-
-    const newTask: Task = {
-      id: count,
-      title: inputTitle,
-      done: false
-    }
-
-    setTasks([...tasks, newTask])
-    setInputTitle('')
-  }
 
   return (
     <div className="App">
-      <div className="inputForm">
-        <input
-          type="text"
-          className="input"
-          value={inputTitle}
-          onChange={handleInputChange}
-        />
-        <button onClick={handleSubmit}>Submit</button>
-      </div>
+      <TaskInputForm tasks={tasks} setTasks={setTasks} />
+      <TaskList tasks={tasks} setTasks={setTasks} />
     </div>
   )
 }
